Change Auto-download settings' title (#2966)
* Change auto-download setting's title

The update don't get automatically downloaded when the user is connected over an metered connection.
This is something the user should known about. Because of not have to be afraid on having extra costs and to not being confused if it not works on metered connections.

* auto-download title (old settings)

Did the same change also for the old web-settings.