Adding Communications section to documentations
Summary of the Pull Request

What is this about: 
After involving myself with PowerToys, I found that it would be really helpful if the PowerToys documentation had a section for communication. More specifically, something such as Slack or Discord as it will dramatically help improve communication and collaboration throughout all contributors. As I have discussed this in a issue before, some individuals love the idea of creating such forums as it will reduce the issues list from a lot of basic questions and problems that do not require the whole community. 

What is included in tthe PR:
Add documentation to the README with a "Communications" section which will include links to Slack/Discord or a relevant 3rd party app. 

How does someone test / validate:
Can be validated by project holders as the communication channels will be created by them. All links will be listed as TBA until further notice from project holders.