[ImageResizer] Fix blurry images in some cases
Fixes #10675

In some cases a resized image becomes very blurry.

Loading the input image file into cache in BitmapDecoder resolves the blurriness.

---

I tried to find the technical reason why caching resolves it, or why it is blurry in the first place, but was unable to locate it.

The issue seems to occur only on specific images. I was not able to identify a strong parameter that would cause it.

Nevertheless, with the images I will attach to the PR, I hope to make an evident case for the factual resolution of the issue. And maybe or hopefully someone with more knowledge about the respective used APIs will be able to provide the factual, technical reason for it.

---

`ResizeOperation` uses `BitmapDecoder` to load the image file. In main the decoder is instantiated with `BitmapCacheOption.None`. Changing it to `BitmapCacheOption.OnLoad` resolves the issue.

The pass-along of image data is as follows:

1. create decoder
2. decoder -> image file metadata
3. decoder -> `BitmapFrame` -> `TransformedBitmap` -> `BitmapFrame` -> encoder (`Frames.Add()`) -> `encoder.Save()`

---

[transformdoc] mentions the following about loaded cached image data potentially mismatching target pixel size, which made me wonder whether the no-cache delayed reading in combination with the transformation leads to reduced-size-reading. I was not able to confirm it. Given that most images I tested did not have a problem, I don't think it's a plausible cause.

> If you don't do this, the application will cache the image as though it were rendered as its normal size rather than just the size that is displayed.

[transformdoc]: https://learn.microsoft.com/en-us/dotnet/desktop/wpf/graphics-multimedia/how-to-apply-a-transform-to-a-bitmapimage

---

For multiple test images, any differences were hard to spot, they did not result in blurriness. The image I had issues with in the past being very colorful made me wonder whether amount of color information could make the difference. But with the technical context, that would only make sense if the encoder caching somehow handled color information differently, which seems unlikely too.