Removed sections from the spec, prototype (#129)
As discussed. 

I'm getting rid of  ``ISection[]`` and replacing that with just a `String Section { get; }`[^1] property on `ListItem`'s themselves. 
 
I didn't want to bother wasting engineering resurrecting the `SectionListView` thing in the prototype without `ISection`s, so I'll just leave that for us to implement The Right Way in the MVVM exe. 
 
This should keep the prototype running, and keep us moving forward

[^1]: we discussed making this an object. I didn't here, but we probably can in the future.