Implement the "explore" page from mastodon (#132)
Actually implements the mastodon extension. Open APIs are great. My baby sleeping for four hours is even better. 

This sample has a list of posts. Each item has Details, which are the post bodies. viewing the post opens a form page with the whole thread of that post and any replies on it. 

* closes #101 
* closes #79 
* Also removes the space for the `HeroImage` if we don't need it. 
* Fixes an issue where forms straight up didn't work (they weren't loading on the main thread)
* Adds a helper `OpenUrl` command (does what it says on the tin)

This extension covers a lot of concepts so I think it's a valuable test. I'm hoping to do auth for allowing you to post/comment/favorite/boost/etc in the future. 

Also helped understand what we need for pagination. 

Also revealed that ItemsRepeaters are cursed.