Create an `IExtensionHost` class, to enable logging (#137)
Allow extensions to send status messages to the host app.

As an example, you can now:

```cs
ExtensionHost.LogMessage(new LogMessage() { Message = "This is a test, from the HackerNews sample" });
```

And have that appear in the host apps debug console. 

Closes #134 
Closes #118 

Should also make resolving #87 and #135 easier. 

We don't have the design of these status messages totally planned out yet. But this should at least make it possible. For those in the back, **this does not actually display these messages anywhere in the host UX currently**. 

-----

Other information that's currently included:
* The host app's HWND. Way too many times does the Windows API need an HWND, which extensions won't have, but the host app does. 
* The `LanguageOverride`. This is in case we ever want to support `PrimaryLanguageOverride` in the host app.