Fix a layout cycle crash in the prototype (#156)
Resurrects the Pokedex extension I was working on before Luca was born.

This is a simple sample of a list with a _lot_ of items, and tags tags tags. It helps to repro a E_LAYOUT_CYCLE that we were seeing. 

This also fixes that layout cycle, by removing that unneeded `StackPanel`


Fixes #WaitItLooksLikeINeverFiledThis