Remaining spec details for the extension lifecycle (#160)
Cleans up a huge remaining TODO section from the spec about the extension lifecycle. 

This documents affordances for extensions to be lazy-loaded, so that we don't always load every extension process. Extensions can also opt-out of this behavior. 

This one doesn't actually have code associated with it. It sounded like a lot to write, for code that would eventually just be tossed next week. 

It does, however, have a drive-by fix for the spongebot extension

Closes #81