Add aliases to the protoype; improve calculator command (#167)
One final yolo for the prototype app, just to see how it feels. 

This adds very rudimentary alias support to the prototype app. I only hardcoded a few of them - calculator, registry, windows settings, and my own virtual desktop switcher. The aliases are powered by the ICommand.Id, added in #160 

Now that we have the alias for the calculator though, it needed to be a lot better. I did that too. Now it saves your history (per launch session), and you can copy the results to the clipboard. And also use TextToSuggest to fill in the current result. And also if you select an old result, it'll plop it back into the search text.