ListItems, ContextItems, and the TopLevel are `ICommandItem`s (#164)
This PR targets #160. A lot of the ideas that led here were written down there first. Especially the "stub" command bits, which precipitated the value for an interface that matched what the top-level commands can really return. 
 
The TL;DR of this PR is found here:
https://github.com/zadjii-msft/PowerToys/issues/153#issuecomment-2486367825

* [x] Closes #153 
* [x] Closes #109 


Shockingly, the code change is kinda small. It's mostly just careful find/replace's. 

Fallback handlers took the biggest delta, but even they work again in the prototype.