ListViewModel Loading Pattern (#180)
On top of #175

This applies the same loading setup between the View and the ViewModel that we had for the ShellPage/ShellViewModel to the ListPage/ListViewModel.

We could abstract this in the View a bit to make work across different combinations of page types and viewmodels later, so they're all consistent, e.g. IInitializableViewModel interface for the `InitializeCommand`.