Adding extensions SDK nuget generation (#184)
## Summary of the Pull Request
Packing the SDK into a nuget package and creating the build scripts and pipeline to build it.
Pipeline: will auto build the SDK and publish it to ADO artifacts in that build.
Script: Devs can build nuget locally by running BuildSDKHelper.ps1.  This will build and pack the SDK nuget and place it in src\modules\cmdpal\extensionsdk\_build.  This directory has been added to the nuget.config as well.