Windows.UI.Color considered harmful (#196)
_targets #192_

Adds our own struct to the API for passing colors around. This is lifted straight out of the Terminal, which had to deal with this quite a lot. 

Since this color is a struct (so that it trivially marshals across the ABI), it can't be _null_. So we also added `OptionalColor`, which is explicitly "Maybe a color, maybe not". We were bit enough times by trying to have a special secret `INVALID_COLOR` value that it ultimately made sense just to put another bool in there. 

While I'm touching the SDK and breaking everyone, I figured I'd also do the `s/Loading/IsLoading/g` thing now too. 

---------

Co-authored-by: Mike Griese <zadjii@gmail.com>