Load icons again (#208)
This adds a magic helper to load icons for us. Any time you want an icon, just do this:

```xaml
<Border Width="16"
        Height="16"
        Margin="4,4,4,4">
    <Interactivity:Interaction.Behaviors>
        <cmdpalUI:LoadIconBehavior Source="{x:Bind ViewModel.PrimaryAction.Icon, Mode=OneWay}"/>
    </Interactivity:Interaction.Behaviors>
</Border>
```

And that'll magically give us a border filled with the icon, and updating with the binding.

I believe it'll also work with `IRandomAccessStreamReference`s, but I didn't actually test that with #151 yet. 

I didn't actually implement the "caching" bit of this yet. That'll involve doing some locking per-key inside the factory and I didn't want to futz with that in this initial PR to restore icons

---------

Co-authored-by: Mike Griese <zadjii@gmail.com>