Add Dynamic lists and Apps on the main page again (#224)
Just to prove that they work. 

This adds the filtering and weighting we were using before to TRA. Lists all get updated in-place, so that the ListView won't flicker like crazy as items get added. And for lists where items get inserted in the middle, we'll update smoothly. 

* 4ad92b64ff429b6959285b6cf88d0feacb643be9...642d98d0571b5d6e03a6faa431c00ede27ce4b24 is what adds support for dynamic lists.
* 09b63d6 I moved some files around to make more sense.
* 09b63d6...2016bcb is where I added apps back onto the main page.

When the `MainListPage` gets a filter string, it will filter both top-level commands and the apps from the `AllAppsPage`. 

I also made the MainListPage show us that it's loading, by setting `IsLoading = appsLoading || extensionsLoading` (more or less).

Does NOT add support for updating fallback items quite yet. One step at a time. 

---------

Co-authored-by: Mike Griese <zadjii@gmail.com>