Introduce an `Id` property for command providers (#269)
Originally a part of #264 

I need command providers to have an ID, so that we can differentiate between multiple providers in the same package.

It's really only relevant for _us_, but I don't know what the future has in store for us. 

Also includes a minor refactoring that combines a couple of the built-ins into a single provider. Reload and quit aren't really separate from _the app itself_.