Fix all the broken form pages (#266)
Random nits, part the second

TRA is particular about form pages. It'll actually break itself if an extension throws an exception (which it should do). The POC was more forgiving (but technically incorrect)

Well, turns out, all the form pages we had did not give an F and just threw exceptions. They shouldn't! That's bad!