Change single/double click behavior on list items (#259)
This PR adjusts the UX for single-clicking / double-clicking on list items. With this change: 

- A single-click on a list item will navigate the user to that list item. It will not automatically invoke the associated command.
- A double-click on a list item will invoke the command for the list item.

It is a really small code change, _but_ I have a suspicion that there are cleaner ways to handle what I'm doing. Since it is so small, I figured just creating the PR and getting quick @zadjii-msft feedback would do the trick.