Add some padding for font icons (#265)
Random nits, part the first.

Emoji render just a little too big currently, and they get clipped. 

This PR fixes that, by making emoji icons have just a bit of negative padding, to give them space to not get clipped. 

Is also replaces the icons on the Services page, because the icon on the list items was the icon of the thing to do to the service (i.e. running services would have a "stop" icon, which is insane). Of course, this made the emoji problem much more apparent.