When an extension item fails to load, replace it with an error item (#293)
Right now, the main page really explodes if an extension dies. It really does not like it. 

Instead, this PR changes it so that if we fail to load the properties from an ICommandItem, we won't kill the whole page, we'll just replace that _single_ item with an error item. The item _would_ look like:

![image](https://github.com/user-attachments/assets/c5e74af1-5d79-47db-81f9-ab6ea1448881)


however, we ended up just silently ignoring the item entirely.