Implement Allowing CommandProviders to ItemsChanged themselves (#289)
This is the implementation of what was spec'd in #282.

When we get an `ItemsChanged`, we'll look through the top level commands, slice out all the old ones, then fetch the new ones and stitch them in the same place in the list. 

I've only implemented it for the Bookmarks provider so far. 

Closes #277