Fix the reload extensions command item (#241)
This never got hooked up in TRA. Now it is. 

And it actually does things smartly - it'll dismiss all the old extension instances, then re-populate all of them. Really helpful for testing extensions. You don't need to keep re-launching the palette, you can just reload and presto - we'll release the old package, let the new one deploy, then load the new one. 

It also revealed that we were updating the `.SearchText` from #234 on the UI thread. Oh no! Can't do that - that might be a x-proc call. Very bad. So that fixes this.