Add support for different "branding" for CmdPal (#299)
The meat of this PR is in `Microsoft.CmdPal.UI\CmdPal.Branding.props`. 

**TL;DR**: You can build cmdpal locally and have it not conflict with the selfhost build, allowing you to use the selfhost CmdPal while _also_ building and testing new features in a _separate_ package. 

This PR allows us to have different packages deployed by the build, depending on the "branding" of cmdpal which you're building. It's highly cribbed from the Terminal's concept of branding. 

This is all keyed off the `CommandPaletteBranding` property. This accepts different values:
* `Release` / `Preview` / `Canary`: These are all the same currently. This is the "release" / "stable" branded version
* `Dev` (or omitted, which is the default for local builds): Build the "dev" version of the package

closes #297 


Co-authored-by: Stefan Markovic <stefan@janeasystems.com>