Fix FontIcon padding, again (#308)
When we changed icons to be a pair of IconData in an IconInfo, I forgot this type check.
Oops. Now it works again.

Originally in #265 
Regressed in #286 

(targets #308, which targets #299)

`IconData` and `IconInfo` are unfortunately, not trivially marshallable into the host process. This creates a collection of cases where an app can crash and take the host down, because we'll try to inquire something about the icon. 

It broke reloading when an extension crashed. 
It broke backing out of a crashed extension. 


Previously: #218 
Closes #235