Fix the tag crash (#339)
Most of my notes are in #322. This one was a journey.

The app was doing something re-entrant when we changed the selected item. It had nothing to do with `ScrollIntoView` (though, not calling that would stop the crash).

Seems like that `ItemsList_SelectionChanged` -> `UpdateSelectedItemCommand` -> `WeakReferenceMessenger.Default.Send<ShowDetailsMessage>`, all on the UI thread, seemed to really make XAML mad. And a combination of hopping to the background thread, and debouncing the `ShowDetails` call, seemed to fix it. 

Tags made it worse - presumbaly it made the XAML layout pass take longer, which widened the window for the app to think it was re-entrant

Closes #322